dpdk.spec: Do not avoid array-bounds warning by patching the sources
By changing the array size to work around a spurious warning from gcc,
we run the risk of breaking future code which may rely on the array's
size (using, ex: sizeof). Since this array-bounds failure is the product
of gcc-5.x, and not actually an error here, it can safely be ignored.

We need to keep the -Wformat warning, however, because that will trigger
build errors in some of the PMD builds.

Signed-off-by: Aaron Conole <aconole@redhat.com>
Acked-by: Neil Horman <nhorman@redhat.com>
Acked-by: Panu Matilainen <pmatilai@redhat.com>