dpdk.spec: Allow builds to have the lightweight api-guide as well as the full pdf build
The PDF guides drag a large amount of build requirements with them, and
because of the reliance on the DejaVu texlive package, will not build on
RHEL7. To keep the option of the PDF guide for Fedora, while allowing
RH to build, make this a tunable which Fedora can default 'on'.

Signed-off-by: Aaron Conole <aconole@redhat.com>
Acked-by: Neil Horman <nhorman@redhat.com>
Acked-by: Panu Matilainen <pmatilai@redhat.com>