dpdk-config: Remove the config patch file, and rewrite the config file post `make config'
The config file patch is a bit cumbersome to maintain. Any time the
config files change, it will be required to regenerate this patch
against the new sets of files. Additionally, it is difficult to build
modular enable/disables for these configs down the line.

With this set, create a block for PMD enable/disable. RHEL wants this
enabled, for sure.

Signed-off-by: Aaron Conole <aconole@redhat.com>
Acked-by: Neil Horman <nhorman@redhat.com>
Acked-by: Panu Matilainen <pmatilai@redhat.com>