dpdk.spec: linkage and file overlap
The dpdk-tools and dpdk packages copy the same files into %{_bindir}, which
should be avoided. To fix this, setup an exclude.

Additionally, there's an odd requirement imposed by using dpdk because of
libraries not being linked in properly against the shared objects generated.
To fix this, introduce the dt-needed patch. A version of this has been
accepted upstream for some PMDs, but we will do this for all.

Signed-off-by: Aaron Conole <aconole@redhat.com>